requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-rest-knox==4.2.0
django-taggit==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
